pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-assistant"
version = "0.1.0"
description = "Hybrid local voice assistant for Windows"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Local Assistant" }]
dependencies = [
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "pydantic>=2.9.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.7.0",
  "mypy>=1.12.0",
  "types-requests>=2.32.0",
  "types-PyYAML>=6.0.12",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = ["--import-mode=importlib", "--strict-markers", "-ra"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
